final/backend/requirements.txt
aioredis==1.1.0
asgiref==2.3.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
cachetools==2.1.0
certifi==2018.4.16
channels==2.1.1
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
daphne==2.1.1
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
google-api-core==1.2.0
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
protobuf==3.5.2.post1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pymongo==3.6.1
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
zipstream==1.1.4
zope.interface==4.5.0
